Restore constant incorrectly removed
