+++ src/constants/stories.ts
@@ -1,0 +1,10 @@
+export const LOREM_IPSUM: string[] = [
+  "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Curabitur congue, libero et imperdiet faucibus, massa neque pellentesque augue, eget egestas lorem ligula id nisi. Fusce pulvinar sem urna, id pretium massa venenatis non. In elementum gravida purus. Proin nec eleifend ligula. Suspendisse potenti. Mauris eu varius massa. Fusce ante ipsum, malesuada vitae massa non, volutpat dictum massa. Duis mauris eros, laoreet et sodales a, blandit dapibus lacus. Vivamus aliquam pellentesque pharetra.",
+  "Praesent mollis purus sed dolor consectetur mattis. Nam gravida, ipsum nec malesuada tempus, odio urna fermentum sem, ut ultrices nulla libero quis velit. Curabitur odio dolor, laoreet a venenatis et, rhoncus ac leo. Vestibulum a diam eget lorem condimentum pharetra. Integer feugiat congue elit. Praesent posuere, magna id ornare mollis, mauris nulla ultricies felis, vel semper nibh nisl vel lorem. Cras nec nisi odio. Quisque eros libero, suscipit vel pellentesque eu, efficitur at mauris. Donec pellentesque ut massa ut egestas. Nulla dictum massa non lorem maximus ornare. Nam molestie libero vel arcu blandit lobortis.",
+  "Pellentesque tempor posuere egestas. Nam placerat ex in tellus varius, vitae varius massa ultrices. Sed egestas nunc nec lacus egestas pellentesque. Cras urna ex, euismod sed magna ac, euismod pretium eros. Nam auctor efficitur quam, ac auctor orci semper a. Aenean id enim ipsum. Donec eget viverra nisi.",
+  "Pellentesque et ante vel neque congue euismod. Maecenas et congue nisl, ut rhoncus ex. Sed et urna scelerisque, suscipit erat eu, semper lorem. Proin ornare gravida dapibus. Vestibulum finibus congue tellus, nec euismod felis vulputate non. Donec ac lacinia massa. Cras nec lorem pulvinar, facilisis risus vitae, dapibus elit. Ut erat leo, feugiat id egestas sed, commodo nec neque. Cras pellentesque sapien sit amet scelerisque malesuada. Curabitur nibh tortor, suscipit at metus non, pulvinar egestas diam. Donec sit amet quam ut quam pretium tincidunt.",
+  "Etiam tortor tortor, laoreet nec porttitor quis, vulputate eu risus. Donec vestibulum maximus ultrices. Aenean eget tellus sapien. Maecenas consequat odio et metus cursus scelerisque. Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas. Morbi gravida erat a sem accumsan condimentum. Curabitur massa quam, pellentesque in urna sed, iaculis ultrices orci. Aliquam et dui tortor. Donec mi erat, sollicitudin ut neque vitae, eleifend finibus mi. Fusce justo neque, sollicitudin tempus auctor in, sodales condimentum metus. Suspendisse imperdiet hendrerit elit eu dictum. Mauris vel malesuada eros. Praesent a porttitor urna. Vestibulum eu venenatis leo.",
+  "Nunc semper, eros commodo sollicitudin scelerisque, odio tellus molestie augue, quis consequat neque velit a risus. Duis a nunc et lorem semper bibendum. Sed libero libero, luctus sit amet est sit amet, mattis pretium leo. Sed eu lacus libero. Mauris a mi viverra, bibendum dui in, interdum metus. Cras a ligula in enim bibendum ornare. Proin interdum lobortis nulla, sed commodo ex. In a felis gravida turpis ornare egestas. Mauris rhoncus, mauris et commodo mollis, erat mauris vestibulum felis, at accumsan lectus nibh a velit.",
+  "Integer ultricies pulvinar imperdiet. Suspendisse non laoreet tortor. Aliquam erat volutpat. Maecenas a sem placerat, vulputate ante vitae, rutrum ex. Donec quis purus sit amet sem mollis porta. Integer hendrerit condimentum odio a varius. In vitae efficitur justo. Phasellus pharetra feugiat ligula, sit amet sagittis diam aliquet pulvinar. Cras vel enim vitae lectus ornare mollis.",
+  "Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas. Morbi at hendrerit risus, a lobortis dolor. Interdum et malesuada fames ac ante ipsum primis in faucibus. Sed porta odio neque, at facilisis justo congue eget. Mauris eu malesuada eros. Pellentesque auctor sagittis tortor, sit amet aliquet neque malesuada vitae. Nullam lacinia dui mollis lectus auctor ullamcorper. Maecenas mattis diam consectetur volutpat iaculis. Nam in eros condimentum ipsum feugiat posuere. Sed id augue nec ante eleifend pulvinar. Duis vitae accumsan tellus, id tristique est.",
+];
